+++ web-client/src/tests/journey/petitionsClerkRecordsPayGovId.js
@@ -1,0 +1,21 @@
+import caseDetailHelper from '../../presenter/computeds/caseDetailHelper';
+
+export default (test, runCompute) => {
+  return it('Petitions clerk records pay.gov ID', async () => {
+    await test.runSequence('updateCaseValueSequence', {
+      key: 'payGovId',
+      value: '123',
+    });
+    await test.runSequence('submitUpdateCaseSequence');
+    test.setState('caseDetail', {});
+    await test.runSequence('gotoCaseDetailSequence', {
+      docketNumber: test.docketNumber,
+    });
+    expect(test.getState('caseDetail.payGovId')).toEqual('123');
+
+    const helper = runCompute(caseDetailHelper, {
+      state: test.getState(),
+    });
+    expect(helper.showPaymentRecord).toEqual(true);
+  });
+};
