+++ web-client/src/tests/journey/respondentLogIn.js
@@ -1,0 +1,10 @@
+export default test => {
+  return it('Respondent logs in', async () => {
+    test.setState('user', {
+      name: 'IRS Attorney',
+      role: 'respondent',
+      token: 'respondent',
+      userId: 'respondent',
+    });
+  });
+};
