+++ web-client/src/tests/journey/respondentViewsCaseDetail.js
@@ -1,0 +1,12 @@
+export default test => {
+  return it('Respondent views case detail', async () => {
+    test.setState('caseDetail', {});
+    await test.runSequence('gotoCaseDetailSequence', {
+      docketNumber: test.docketNumber,
+    });
+    expect(test.getState('currentPage')).toEqual('CaseDetailRespondent');
+    expect(test.getState('caseDetail.docketNumber')).toEqual(test.docketNumber);
+    expect(test.getState('caseDetail.status')).toEqual('general');
+    expect(test.getState('caseDetail.documents').length).toEqual(3);
+  });
+};
