+++ web-client/src/tests/journey/respondentViewsDashboard.js
@@ -1,0 +1,7 @@
+export default test => {
+  return it('Respondent views dashboard', async () => {
+    await test.runSequence('gotoDashboardSequence');
+    expect(test.getState('currentPage')).toEqual('DashboardRespondent');
+    expect(test.getState('cases').length).toBeGreaterThan(0);
+  });
+};
